added user registration and restricted pages

+++ stumblr/base/views.py
@@ -1,16 +1,22 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from django.db.models import Q
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib.auth import login,logout,authenticate
+from django.contrib.auth.forms import UserCreationForm
 
 
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -23,12 +29,27 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request,'Username or Password does not exist')
-    context = {}
+    context = {'page':page}
     return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid:
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during registration')
+    return render(request,'base/login_register.html',{'form':form})
 
 
 
@@ -50,6 +71,8 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -60,9 +83,12 @@ def createRoom(request):
     context ={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to do that!')
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -71,8 +97,11 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to do that!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
